Te/distribution components (#30)
* update bus and integrate into distribution
* removing capacitor base class and integrating into phase_capacitor
* removing transmission base class (unused)
* updating to use DistributionComponent rather than belongs_to
* setting bus names to be specific for the component
* removing transmission elements from gdm
* updating version to 0.3.0
