+++ src/gdm/distribution/components/distribution_bus.py
@@ -1,19 +1,20 @@
 """ This module contains interface for Distribution Bus."""
 
-from typing import Annotated
+from typing import Annotated, Optional
 
 from infrasys import Location
 from pydantic import Field
 
 from gdm.distribution.distribution_enum import LimitType, Phase, VoltageTypes
 from gdm.distribution.components.distribution_component import DistributionComponent
-from gdm.distribution.distribution_common import BELONG_TO_TYPE
+from gdm.distribution.components.distribution_feeder import DistributionFeeder
+from gdm.distribution.components.distribution_substation import DistributionSubstation
 from gdm.distribution.limitset import VoltageLimitSet
 from gdm.quantities import PositiveVoltage
-from gdm.bus import PowerSystemBus
+from gdm.constants import PINT_SCHEMA
 
 
-class DistributionBus(PowerSystemBus):
+class DistributionBus(DistributionComponent):
     """Interface for distribution bus.
 
     Examples:
@@ -22,21 +23,30 @@ class DistributionBus(PowerSystemBus):
     """
 
     voltage_type: Annotated[VoltageTypes, Field(..., description="Voltage types for buses.")]
-    belongs_to: BELONG_TO_TYPE
     phases: Annotated[list[Phase], Field(..., description="List of phases for this bus.")]
     voltagelimits: Annotated[
         list[VoltageLimitSet],
         Field([], description="List of voltage limit sets for this bus."),
+    ]
+    nominal_voltage: Annotated[
+        PositiveVoltage,
+        PINT_SCHEMA,
+        Field(..., description="Nominal voltage for this bus."),
+    ]
+    coordinate: Annotated[
+        Optional[Location],
+        Field(None, description="Coordinate for the power system bus."),
     ]
 
     @classmethod
     def example(cls) -> "DistributionBus":
         return DistributionBus(
             voltage_type=VoltageTypes.LINE_TO_LINE,
-            belongs_to=DistributionComponent.example(),
             phases=[Phase.A, Phase.B, Phase.C],
             nominal_voltage=PositiveVoltage(400, "volt"),
             name="DistBus1",
+            substation=DistributionSubstation.example(),
+            feeder=DistributionFeeder.example(),
             voltagelimits=[
                 VoltageLimitSet(
                     limit_type=LimitType.MIN, value=PositiveVoltage(400 * 0.9, "volt")
