+++ src/gdm/distribution/components/distribution_component.py
@@ -13,7 +13,6 @@ from gdm.distribution.components.distribution_substation import DistributionSubs
 class DistributionComponent(Component):
     """Interface for simple distribution component."""
 
-    name: Annotated[str, Field("", description="Name of the component.")]
     substation: Annotated[
         Optional[DistributionSubstation], Field(None, description="Name of the substation.")
     ]
@@ -22,5 +21,7 @@ class DistributionComponent(Component):
     @classmethod
     def example(cls) -> "DistributionComponent":
         return DistributionComponent(
-            substation=DistributionSubstation.example(), feeder=DistributionFeeder.example()
+            name="BaseComponent",
+            substation=DistributionSubstation.example(),
+            feeder=DistributionFeeder.example()
         )
