+++ src/gdm/distribution/components/distribution_load.py
@@ -3,16 +3,17 @@
 from typing import Annotated
 
 from pydantic import model_validator, Field
-from infrasys import Component
 
 from gdm.distribution.distribution_enum import Phase
-from gdm.distribution.distribution_common import BELONG_TO_TYPE
 from gdm.distribution.components.distribution_bus import DistributionBus
+from gdm.distribution.components.distribution_component import DistributionComponent
+from gdm.distribution.components.distribution_feeder import DistributionFeeder
+from gdm.distribution.components.distribution_substation import DistributionSubstation
 from gdm.quantities import PositiveVoltage
 from gdm.distribution.equipment.load_equipment import LoadEquipment
 
 
-class DistributionLoad(Component):
+class DistributionLoad(DistributionComponent):
     """Interface for distribution load."""
 
     bus: Annotated[
@@ -22,7 +23,6 @@ class DistributionLoad(Component):
             description="Distribution bus to which this load is connected to.",
         ),
     ]
-    belongs_to: BELONG_TO_TYPE
     phases: Annotated[
         list[Phase],
         Field(..., description="Phases to which this load is connected to."),
@@ -51,10 +51,14 @@ class DistributionLoad(Component):
             name="DistributionLoad1",
             bus=DistributionBus(
                 voltage_type="line-to-ground",
-                name="Bus1",
+                name="Load-DistBus1",
                 phases=[Phase.A, Phase.B, Phase.C],
+                substation=DistributionSubstation.example(),
+                feeder=DistributionFeeder.example(),
                 nominal_voltage=PositiveVoltage(0.4, "kilovolt"),
             ),
+            substation=DistributionSubstation.example(),
+            feeder=DistributionFeeder.example(),
             phases=[Phase.A, Phase.B, Phase.C],
             equipment=LoadEquipment.example(),
         )
