+++ src/gdm/distribution/components/distribution_solar.py
@@ -2,13 +2,14 @@
 
 from typing import Annotated
 
-from infrasys import Component
 from pydantic import Field
 
 from gdm.quantities import PositiveVoltage
 from gdm.distribution.distribution_enum import Phase
-from gdm.distribution.distribution_common import BELONG_TO_TYPE
 from gdm.distribution.components.distribution_bus import DistributionBus
+from gdm.distribution.components.distribution_component import DistributionComponent
+from gdm.distribution.components.distribution_feeder import DistributionFeeder
+from gdm.distribution.components.distribution_substation import DistributionSubstation
 from gdm.distribution.controllers.distribution_inverter_controller import (
     InverterController,
     VoltVarInverterController,
@@ -16,7 +17,7 @@ from gdm.distribution.controllers.distribution_inverter_controller import (
 from gdm.distribution.equipment.solar_equipment import SolarEquipment
 
 
-class DistributionSolar(Component):
+class DistributionSolar(DistributionComponent):
     """Interface for Solar PV system in distribution system models."""
 
     bus: Annotated[
@@ -26,7 +27,6 @@ class DistributionSolar(Component):
             description="Distribution bus to which this solar array is connected to.",
         ),
     ]
-    belongs_to: BELONG_TO_TYPE
     phases: Annotated[
         list[Phase],
         Field(
@@ -53,10 +53,14 @@ class DistributionSolar(Component):
             name="pv1",
             bus=DistributionBus(
                 voltage_type="line-to-ground",
-                name="Bus1",
+                name="Solar-DistBus1",
                 nominal_voltage=PositiveVoltage(400, "volt"),
                 phases=[Phase.A, Phase.B, Phase.C],
+                substation=DistributionSubstation.example(),
+                feeder=DistributionFeeder.example(),
             ),
+            substation=DistributionSubstation.example(),
+            feeder=DistributionFeeder.example(),
             phases=[Phase.A, Phase.B, Phase.C],
             equipment=SolarEquipment.example(),
             controller=VoltVarInverterController.example(),
