+++ src/gdm/distribution/components/sequence_impedance_branch.py
@@ -5,6 +5,8 @@ from typing import Annotated
 from pydantic import Field, model_validator
 
 from gdm.distribution.components.distribution_branch import DistributionBranch
+from gdm.distribution.components.distribution_feeder import DistributionFeeder
+from gdm.distribution.components.distribution_substation import DistributionSubstation
 from gdm.distribution.equipment.sequence_impedance_branch_equipment import (
     SequenceImpedanceBranchEquipment,
 )
@@ -34,5 +36,7 @@ class SequenceImpedanceBranch(DistributionBranch):
             length=base_branch.length,
             phases=base_branch.phases,
             name=base_branch.name,
+            substation=DistributionSubstation.example(),
+            feeder=DistributionFeeder.example(),
             equipment=SequenceImpedanceBranchEquipment.example(),
         )
