+++ src/gdm/distribution/equipment/phase_capacitor_equipment.py
@@ -2,14 +2,14 @@
 
 from typing import Annotated
 
-from pydantic import Field
+from pydantic import Field, NonNegativeInt, PositiveInt, model_validator
+from infrasys import Component
 
-from gdm.capacitor import PowerSystemCapacitor
-from gdm.quantities import PositiveResistance, PositiveReactance
+from gdm.quantities import PositiveResistance, PositiveReactance, PositiveReactivePower
 from gdm.constants import PINT_SCHEMA
 
 
-class PhaseCapacitorEquipment(PowerSystemCapacitor):
+class PhaseCapacitorEquipment(Component):
     """Interface for phase capacitor."""
 
     resistance: Annotated[
@@ -28,14 +28,32 @@ class PhaseCapacitorEquipment(PowerSystemCapacitor):
             description="Positive reactance for the capacitor.",
         ),
     ]
+    rated_capacity: Annotated[
+        PositiveReactivePower,
+        PINT_SCHEMA,
+        Field(..., description="Capacity of this capacitor."),
+    ]
+    num_banks_on: Annotated[
+        NonNegativeInt, Field(..., description="Number of banks currently on.")
+    ]
+    num_banks: Annotated[PositiveInt, Field(1, description="Number of banks in the capacitor.")]
+
+    @model_validator(mode="after")
+    def validate_fields(self) -> "PhaseCapacitorEquipment":
+        """Custom validator for fields."""
+        if self.num_banks < self.num_banks_on:
+            msg = f"Status {self.num_banks_on} must be less than or equal"
+            f"to number of banks. {self.num_banks}"
+            raise ValueError(msg)
+
+        return self
 
     @classmethod
     def example(cls) -> "PhaseCapacitorEquipment":
         """Example for phase capacitor equipment."""
-        base_cap = PowerSystemCapacitor.example()
         return PhaseCapacitorEquipment(
-            name=base_cap.name,
-            rated_capacity=base_cap.rated_capacity,
-            num_banks=base_cap.num_banks,
-            num_banks_on=base_cap.num_banks_on,
+            name="Phase-Cap-1",
+            rated_capacity=PositiveReactivePower(200,"kvar"),
+            num_banks=1,
+            num_banks_on=1,
         )
