--- src/gdm/transformer.py
@@ -1,7 +1,0 @@
-"""This module contains base power system transformer interface."""
-
-from infrasys import Component
-
-
-class PowerSystemTransformer(Component):
-    """Interface for power system transformer."""
