--- src/gdm/transmission/transmission_substation.py
@@ -1,10 +1,0 @@
-"""This class contains interface for transmission substation."""
-
-from infrasys import Component
-from gdm.transmission.transmission_bus import TransmissionBus
-
-
-class TransmissionSubstation(Component):
-    """Interface for transmission substation."""
-
-    bus: TransmissionBus
