+++ src/gdm/version.py
@@ -3,7 +3,7 @@ import subprocess
 import platform
 import sys
 
-VERSION = "0.2.0"
+VERSION = "0.3.0"
 
 
 def is_git_repo(dir: Path) -> bool:
